Added time field in the UI

+++ app/src/main/java/com/github/welshk/ukko/data/models/WeatherDetails.kt
@@ -19,7 +19,7 @@ data class WeatherDetails(
    @SerializedName("clouds")
     var clouds: Clouds? = null,
    @SerializedName("dt")
-    var dt: Int,
+    var dt: Long,
    @SerializedName("sys")
     var sys: Sys? = null,
    @SerializedName("timezone")
